:bug: 主键修改为 Long 类型遗漏部分提交 (#254)

+++ ballcat-system/ballcat-system-model/src/main/java/com/hccake/ballcat/system/model/dto/SysUserDTO.java
@@ -92,7 +92,7 @@ public class SysUserDTO {
 	 * 组织机构ID
 	 */
 	@Schema(title = "组织机构ID")
-	private Integer organizationId;
+	private Long organizationId;
 
 	/**
 	 * 角色标识列表
